fix:update file structure for Impact
