+++ src/loaders/index.ts
@@ -1,18 +1,12 @@
-import { fetcher } from "../utils";
 import { LoaderFunction } from "react-router-dom";
 import {
   icuCapacityMeterLoader,
   progressBarLoader,
   testPositiveLoader,
 } from "./feedback";
-import {
-  vacRateLoader,
-  deathRateLoader,
-} from "./group11";
-
-import {
-  lollipopChartLoader,
-} from "./group5";
+import { vacRateLoader, deathRateLoader } from "./group11";
+import { lollipopChartLoader } from "./group5";
+import { treeMapLoader, sunburstLoader } from "./impact";
 
 export const feedbackLoader = (async (): Promise<any> => {
   const progresssBarData = await progressBarLoader();
@@ -29,84 +23,11 @@ export const group11Loader = (async (): Promise<any> => {
 
 export const group5Loader = (async (): Promise<any> => {
   const lollipopData = await lollipopChartLoader();
-  return { lollipopData};
+  return { lollipopData };
 }) satisfies LoaderFunction;
 
-export const sunburstLoader = (async (): Promise<any> => {
-  const districtData = await fetcher("vaccination/vax_district.csv");
-  const chartData = districtData.reduce(
-    (result: any, row: any) => {
-      const { state, district, daily_partial } = row;
-
-      if (!state || !district) {
-        return result;
-      }
-
-      let stateNode = result.children.find((node: any) => node.name === state);
-
-      if (!stateNode) {
-        stateNode = { name: state, children: [] };
-        result.children.push(stateNode);
-      }
-
-      const districtNode = stateNode.children.find(
-        (node: any) => node.name === district
-      );
-
-      if (districtNode) {
-        districtNode.value += parseInt(daily_partial, 10);
-      } else {
-        stateNode.children.push({
-          name: district,
-          value: parseInt(daily_partial, 10),
-        });
-      }
-
-      return result;
-    },
-    { name: "home", children: [] }
-  );
-
-  return chartData;
-}) satisfies LoaderFunction;
-
-export const treeMapLoader = (async (): Promise<any[]> => {
-  const districtData = await fetcher("vaccination/vax_district.csv");
-  const treeMapData: any[] = (districtData as any[]).reduce(
-    (result: any, row: any) => {
-      const { state, district, daily_partial } = row;
-
-      if (!state || !district) {
-        return result;
-      }
-
-      let stateNode = result.find((node: any) => node.name === state);
-
-      if (!stateNode) {
-        stateNode = { name: state, children: [] };
-        result.push(stateNode);
-      }
-
-      const districtNode = stateNode.children.find(
-        (node: any) => node.name === district
-      );
-
-      if (districtNode) {
-        districtNode.value += parseInt(daily_partial, 10);
-        if (districtNode.value > 500000) {
-          districtNode.showLabel = true;
-        }
-      } else {
-        stateNode.children.push({
-          name: district,
-          value: parseInt(daily_partial, 10),
-        });
-      }
-
-      return result;
-    },
-    []
-  );
-
-  return treeMapData;
+export const impactLoader = (async (): Promise<any> => {
+  const treeMapData = await treeMapLoader();
+  const sunburstData = await sunburstLoader();
+  return { treeMapData, sunburstData };
 }) satisfies LoaderFunction;
