+++ src/pages/index.ts
@@ -1,7 +1,6 @@
 export { default as Overview } from "./Overview";
 export { default as Feedback } from "./Feedback";
-export { default as SunburstPage } from "./SunburstPage";
-export { default as TreemapPage } from "./TreeMapPage";
+export { default as ImpactPage } from "./Impact";
 export { default as Group11Page } from "./Group11";
 export { default as Group5Page } from "./Group5";
